Add files via upload
 Guided Project - PREDICTIVE MAINTENANCE OF ENGINES - Bhavna Ajit Sharma - S P Jain School of Global Management
This particular project is done by Pre-Processing the data, Analyzing the pre-processed data, training the machine with pre-processed data using an appropriate machine learning algorithm, saving the model and its dependencies, and lastly building a Web application using flask that integrates with the built model. This approach based on past data is the one we used to build this project.